refactor(handler) split header_filter to its own file
### Summary

Removes `header_filter` code from `handler.lua` and moves it to
`header_filter.lua`.
